# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tinder App API",
        "description": "API for Tinder-like dating app MVP",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/v1/recommendations": {
            "get": {
                "tags": [
                    "Recommendations"
                ],
                "summary": "Get list of recommended people",
                "operationId": "b0e8cb534e8ff931e4f834d2842695ca",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of recommended people",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "string",
                                            "example": "SUCCESS"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "List of recommended people"
                                        },
                                        "data": {
                                            "properties": {
                                                "items": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Person"
                                                    }
                                                },
                                                "meta": {
                                                    "properties": {
                                                        "page": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "per_page": {
                                                            "type": "integer",
                                                            "example": 20
                                                        },
                                                        "total": {
                                                            "type": "integer",
                                                            "example": 85
                                                        },
                                                        "total_pages": {
                                                            "type": "integer",
                                                            "example": 5
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "UserIdHeader": []
                    }
                ]
            }
        },
        "/v1/people/{id}": {
            "get": {
                "tags": [
                    "Recommendations"
                ],
                "summary": "Get person detail",
                "operationId": "ae393d239349719e6055e61beff17a2d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Person ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Person detail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "string",
                                            "example": "SUCCESS"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Person detail"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Person"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Person not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "string",
                                            "example": "PERSON_NOT_FOUND"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Person not found."
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "UserIdHeader": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Person": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 116
                    },
                    "name": {
                        "type": "string",
                        "example": "Ayu"
                    },
                    "age": {
                        "type": "integer",
                        "example": 26
                    },
                    "location": {
                        "type": "string",
                        "example": "Jakarta"
                    },
                    "pictures": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "url": {
                                    "type": "string",
                                    "example": "https://.../ayu1.jpg"
                                },
                                "order": {
                                    "type": "integer",
                                    "example": 1
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "UserIdHeader": {
                "type": "apiKey",
                "description": "User ID for testing (UUID format)",
                "name": "X-User-Id",
                "in": "header"
            }
        }
    },
...
}